Relocate shared validation UI to Create/Components and refine WSDL create step
Move validationErrorUtils and ValidationResults out of Create/OpenAPI/Steps
into the shared Create/Components directory so both the OpenAPI and WSDL
create flows consume them from a common location, and update the imports in
ProvideOpenAPI and ProvideWSDL accordingly.

WSDL Provide step:
- Make the input row full-width and drop the container spacing so the drop
  zone and the validation results panel match the OpenAPI step layout.
- Reset the validation state (results panel, inline error and selected input)
  when the Input Type or Implementation Type changes, matching the OpenAPI
  flow, so a stale error no longer lingers after switching modes.
