+++ portals/publisher/src/main/webapp/source/src/app/components/Apis/Create/Components/validationErrorUtils.ts
